central_terminal/ws_bridge_requirements.txt
aiohttp>=3.9.0
nats-py>=2.7.0
